add RushHourEvent

+++ src/meowtro/eventSystem/holidayEvent/HolidayEvent.java
@@ -3,8 +3,8 @@ import meowtro.eventSystem.Event;
 import meowtro.game.City;
 
 public abstract class HolidayEvent extends Event{
-    protected float growthRate;
-    public HolidayEvent(City city, String happenedTimeString, float growthRate){
+    protected double growthRate;
+    public HolidayEvent(City city, String happenedTimeString, double growthRate){
         super(city, happenedTimeString);
         this.growthRate = growthRate;
     }
